+++ apps/vr-tests/src/stories/MessageBar.stories.tsx
@@ -22,14 +22,14 @@ storiesOf('MessageBar', module)
       {story()}
     </Screener>
   ))
-  .add('Root', () => (<MessageBar>Info/default message bar. {link}</MessageBar>))
-  .add('Root dismiss', () => (
-    <MessageBar onDismiss={noop}>Info/default message bar. {link}</MessageBar>))
-  .add('Root dismiss single line', () => (
+  .addStory('Root', () => (<MessageBar>Info/default message bar. {link}</MessageBar>), { rtl: true })
+  .addStory('Root dismiss', () => (
+    <MessageBar onDismiss={noop}>Info/default message bar. {link}</MessageBar>), { rtl: true })
+  .addStory('Root dismiss single line', () => (
     <MessageBar onDismiss={noop} isMultiline={false}>Info/default message bar. {link}</MessageBar>))
-  .add('Root truncated', () => (
-    <MessageBar truncated={true} isMultiline={false}>Blocked lorem ipsum dolor sit amet, consectetur adipiscing elit. Morbi luctus, purus a lobortis tristique, odio augue pharetra metus, ac placerat nunc mi nec dui. Vestibulum aliquam et nunc semper scelerisque. Curabitur vitae orci nec quam condimentum porttitor et sed lacus. Vivamus ac efficitur leo. Cras faucibus mauris libero, ac placerat erat euismod et. Donec pulvinar commodo odio sit amet faucibus. In hac habitasse platea dictumst. Duis eu ante commodo, condimentum nibh pellentesque, laoreet enim. Fusce massa lorem, ultrices eu mi a, fermentum suscipit magna. Integer porta purus pulvinar, hendrerit felis eget, condimentum mauris. {link}</MessageBar>))
-  .add('Root actions', () => (
+  .addStory('Root truncated', () => (
+    <MessageBar truncated={true} isMultiline={false}>Blocked lorem ipsum dolor sit amet, consectetur adipiscing elit. Morbi luctus, purus a lobortis tristique, odio augue pharetra metus, ac placerat nunc mi nec dui. Vestibulum aliquam et nunc semper scelerisque. Curabitur vitae orci nec quam condimentum porttitor et sed lacus. Vivamus ac efficitur leo. Cras faucibus mauris libero, ac placerat erat euismod et. Donec pulvinar commodo odio sit amet faucibus. In hac habitasse platea dictumst. Duis eu ante commodo, condimentum nibh pellentesque, laoreet enim. Fusce massa lorem, ultrices eu mi a, fermentum suscipit magna. Integer porta purus pulvinar, hendrerit felis eget, condimentum mauris. {link}</MessageBar>), { rtl: true })
+  .addStory('Root actions', () => (
     <MessageBar
       actions={
         <div>
@@ -38,8 +38,8 @@ storiesOf('MessageBar', module)
         </div>
       }
     >Info/default message bar. {link}
-    </MessageBar>))
-  .add('Root actions single line', () => (
+    </MessageBar>), { rtl: true })
+  .addStory('Root actions single line', () => (
     <MessageBar
       isMultiline={false}
       actions={
@@ -49,8 +49,8 @@ storiesOf('MessageBar', module)
         </div>
       }
     >Info/default message bar. {link}
-    </MessageBar>))
-  .add('Root dismiss and action', () => (
+    </MessageBar>), { rtl: true })
+  .addStory('Root dismiss and action', () => (
     <MessageBar
       onDismiss={noop}
       actions={
@@ -60,8 +60,8 @@ storiesOf('MessageBar', module)
         </div>
       }
     >Info/default message bar. {link}
-    </MessageBar>))
-  .add('Root dismiss and action single line', () => (
+    </MessageBar>), { rtl: true })
+  .addStory('Root dismiss and action single line', () => (
     <MessageBar
       isMultiline={false}
       onDismiss={noop}
@@ -72,26 +72,26 @@ storiesOf('MessageBar', module)
         </div>
       }
     >Info/default message bar. {link}
-    </MessageBar>))
-  .add('Root multiline', () => (<MessageBar isMultiline>Info/default message bar. {longText}</MessageBar>))
-  .add('Root overflow', () => (<MessageBar isMultiline={false}>Info/default message bar. {longText}  </MessageBar>))
-  .add('Error', () => (
+    </MessageBar>), { rtl: true })
+  .addStory('Root multiline', () => (<MessageBar isMultiline>Info/default message bar. {longText}</MessageBar>), { rtl: true })
+  .addStory('Root overflow', () => (<MessageBar isMultiline={false}>Info/default message bar. {longText}  </MessageBar>), { rtl: true })
+  .addStory('Error', () => (
     <MessageBar messageBarType={MessageBarType.error}>
       Error message bar. {link}
     </MessageBar>))
-  .add('Blocked', () => (
+  .addStory('Blocked', () => (
     <MessageBar messageBarType={MessageBarType.blocked}>
       Blocked message bar. {link}
     </MessageBar>))
-  .add('Severe Warning', () => (
+  .addStory('Severe Warning', () => (
     <MessageBar messageBarType={MessageBarType.severeWarning}>
       Severe Warning message bar. {link}
     </MessageBar>))
-  .add('Success', () => (
+  .addStory('Success', () => (
     <MessageBar messageBarType={MessageBarType.success}>
       Success message bar. {link}
     </MessageBar>))
-  .add('Warning', () => (
+  .addStory('Warning', () => (
     <MessageBar messageBarType={MessageBarType.warning}>
       Warning message bar. {link}
     </MessageBar>));