+++ packages/dashboard/src/components/Card/Card.tsx
@@ -33,6 +33,7 @@ export class Card extends React.Component<ICardProps, ICardState> {
           cardTitle={cardFrameContent.cardTitle}
           cardDropDownOptions={cardFrameContent.cardDropDownOptions}
           href={cardFrameContent.href}
+          target={cardFrameContent.target}
           disableDrag={disableDrag === undefined ? false : disableDrag}
         >
           <Layout header={header} contentArea={cardContentList} cardSize={this.state.cardSize} actions={actions} />