+++ packages/dashboard/src/components/Card/GridList/GridList.tsx
@@ -1,12 +1,6 @@
 import * as React from 'react';
 import { IGridListProps, IGridListStyles, IGridColumn, IGridRow, GridColumnContentType } from './GridList.types';
-import {
-  DetailsList,
-  IColumn,
-  ColumnActionsMode,
-  DetailsListLayoutMode,
-  ConstrainMode
-} from 'office-ui-fabric-react/lib/DetailsList';
+import { DetailsList, IColumn, ColumnActionsMode, DetailsListLayoutMode, ConstrainMode } from 'office-ui-fabric-react/lib/DetailsList';
 import { Icon } from 'office-ui-fabric-react/lib/Icon';
 import { Image, ImageFit } from 'office-ui-fabric-react/lib/Image';
 import { CheckboxVisibility } from 'office-ui-fabric-react';
@@ -55,7 +49,7 @@ export class GridList extends React.Component<IGridListProps> {
     } else {
       cell = row.c3;
     }
-    const customColorCss = getClassNames(getStyles({ iconColor: cell.iconColor, textColor: cell.textColor }));
+    const customColorCss = getClassNames(getStyles({ iconColor: cell.iconColor, textColor: cell.textColor, boldText: cell.boldText }));
     switch (column.key) {
       case 'facepile':
         return (
@@ -84,10 +78,7 @@ export class GridList extends React.Component<IGridListProps> {
     }
   }
 
-  private _renderActionButton(
-    actionButtonText: string | undefined,
-    onActionLinkClicked: VoidFunction | undefined
-  ): JSX.Element | null {
+  private _renderActionButton(actionButtonText: string | undefined, onActionLinkClicked: VoidFunction | undefined): JSX.Element | null {
     if (actionButtonText !== undefined) {
       return (
         <ActionButton onClick={onActionLinkClicked}>
@@ -114,21 +105,24 @@ export class GridList extends React.Component<IGridListProps> {
           facepileImageSrc: gridRow.c1.facepileImageSrc,
           iconName: gridRow.c1.iconName,
           textColor: gridRow.c1 !== undefined ? gridRow.c1.textColor : undefined,
-          iconColor: gridRow.c1 !== undefined ? gridRow.c1.iconColor : undefined
+          iconColor: gridRow.c1 !== undefined ? gridRow.c1.iconColor : undefined,
+          boldText: gridRow.c1 !== undefined ? gridRow.c1.boldText : undefined
         },
         c2: {
           content: gridRow.c2 !== undefined ? gridRow.c2.content : undefined,
           facepileImageSrc: gridRow.c2 !== undefined ? gridRow.c2.facepileImageSrc : undefined,
           iconName: gridRow.c2 !== undefined ? gridRow.c2.iconName : undefined,
           textColor: gridRow.c1 !== undefined ? gridRow.c1.textColor : undefined,
-          iconColor: gridRow.c1 !== undefined ? gridRow.c1.iconColor : undefined
+          iconColor: gridRow.c1 !== undefined ? gridRow.c1.iconColor : undefined,
+          boldText: gridRow.c1 !== undefined ? gridRow.c1.boldText : undefined
         },
         c3: {
           content: gridRow.c3 !== undefined ? gridRow.c3.content : undefined,
           facepileImageSrc: gridRow.c3 !== undefined ? gridRow.c3.facepileImageSrc : undefined,
           iconName: gridRow.c3 !== undefined ? gridRow.c3.iconName : undefined,
           textColor: gridRow.c1 !== undefined ? gridRow.c1.textColor : undefined,
-          iconColor: gridRow.c1 !== undefined ? gridRow.c1.iconColor : undefined
+          iconColor: gridRow.c1 !== undefined ? gridRow.c1.iconColor : undefined,
+          boldText: gridRow.c1 !== undefined ? gridRow.c1.boldText : undefined
         }
       };
       rows.push(rowItem);