+++ packages/dashboard/src/index.ts
@@ -5,3 +5,5 @@ export * from './components/Recommendation/index';
 export * from './components/MultiCount/index';
 export * from './components/Section/index';
 export * from './utilities/index';
+
+import './version';