+++ packages/experiments/src/components/Stack/Stack.styles.ts
@@ -74,7 +74,10 @@ export const styles = (props: IThemedProps<IStackProps>): IStackStyles => {
         selectors: {
           '> *': {
             margin: `${0.5 * vGap.value}${vGap.unit} ${0.5 * hGap.value}${hGap.unit}`,
-            maxWidth: `calc(100% - ${hGap.value}${hGap.unit})`,
+
+            // extra 2px to account for padding on wrapped Stacks
+            maxWidth: `calc(100% - ${hGap.value}${hGap.unit} - 2px)`,
+
             ...childStyles
           },
           ...commonSelectors