+++ packages/office-ui-fabric-react/src/index.ts
@@ -67,3 +67,5 @@ export * from './TextField';
 export * from './Toggle';
 export * from './Tooltip';
 export * from './Utilities';
+
+import './version';