+++ packages/styling/src/styles/index.ts
@@ -6,14 +6,8 @@ export { getFocusStyle, focusClear } from './getFocusStyle';
 export { hiddenContentStyle } from './hiddenContentStyle';
 export { PulsingBeaconAnimationStyles } from './PulsingBeaconAnimationStyles';
 export { getGlobalClassNames, GlobalClassNames } from './getGlobalClassNames';
-export {
-  ThemeSettingName,
-  getTheme,
-  loadTheme,
-  createTheme,
-  registerOnThemeChangeCallback,
-  removeOnThemeChangeCallback
-} from './theme';
+export * from './scheme';
+export { ThemeSettingName, getTheme, loadTheme, createTheme, registerOnThemeChangeCallback, removeOnThemeChangeCallback } from './theme';
 export * from './CommonStyles';
 export * from './GeneralStyles';
 export * from './getFadedOverflowStyle';