+++ packages/experiments/CHANGELOG.md
@@ -1,6 +1,21 @@
 # Change Log - @uifabric/experiments
 
-This log was last generated on Wed, 19 Sep 2018 12:27:48 GMT and should not be manually modified.
+This log was last generated on Tue, 25 Sep 2018 12:28:12 GMT and should not be manually modified.
+
+## 6.32.2
+Tue, 25 Sep 2018 12:28:12 GMT
+
+### Patches
+
+- Stack: update styles to account for all margin collapsing on wrapped HorizontalStacks
+
+## 6.32.1
+Fri, 21 Sep 2018 14:25:46 GMT
+
+### Patches
+
+- Adding a version stamp file
+- Example modification to highlight scheming capability.
 
 ## 6.32.0
 Wed, 19 Sep 2018 12:27:48 GMT